Add stories to evaluate theme new colors
