+++ newIDE/app/src/stories/index.js
@@ -216,18 +216,6 @@ addDecorator(GDevelopJsInitializerDecorator);
 
 // No i18n in this file
 
-const buildFakeMenuTemplate = () => [
-  {
-    label: 'Option 1',
-    click: action('click option 1'),
-  },
-  { type: 'separator' },
-  {
-    label: 'Option 2',
-    click: action('click option 2'),
-  },
-];
-
 const hotReloadPreviewButtonProps: HotReloadPreviewButtonProps = {
   hasPreviewsRunning: false,
   launchProjectDataOnlyPreview: action('launchProjectDataOnlyPreview'),
@@ -239,182 +227,6 @@ const hotReloadPreviewButtonProps: HotReloadPreviewButtonProps = {
 storiesOf('Welcome', module)
   .addDecorator(muiDecorator)
   .add('to Storybook', () => <Welcome />);
-
-storiesOf('UI Building Blocks/Buttons', module)
-  .addDecorator(paperDecorator)
-  .addDecorator(muiDecorator)
-  .add('default', () => (
-    <ColumnStackLayout>
-      <LineStackLayout noMargin>
-        <Text>Buttons:</Text>
-      </LineStackLayout>
-      <LineStackLayout noMargin>
-        <RaisedButton label="Raised button" onClick={action('onClick')} />
-        <RaisedButton
-          icon={<CloudDownload />}
-          label="Raised button"
-          onClick={action('onClick')}
-        />
-        <RaisedButton
-          label="Primary Raised button"
-          primary
-          onClick={action('onClick')}
-        />
-        <RaisedButton
-          icon={<CloudDownload />}
-          label="Primary Raised button"
-          primary
-          onClick={action('onClick')}
-        />
-      </LineStackLayout>
-      <LineStackLayout noMargin>
-        <FlatButton label="Flat button" onClick={action('onClick')} />
-        <FlatButton
-          leftIcon={<CloudDownload />}
-          label="Flat button"
-          onClick={action('onClick')}
-        />
-        <FlatButton
-          label="Primary Flat button"
-          primary
-          onClick={action('onClick')}
-        />
-        <FlatButton
-          leftIcon={<CloudDownload />}
-          label="Primary Flat button"
-          primary
-          onClick={action('onClick')}
-        />
-      </LineStackLayout>
-      <LineStackLayout noMargin>
-        <TextButton label="Text button" onClick={action('onClick')} />
-        <TextButton
-          icon={<CloudDownload />}
-          label="Text button"
-          onClick={action('onClick')}
-        />
-        <TextButton
-          primary
-          label="Primary Text button"
-          onClick={action('onClick')}
-        />
-        <TextButton
-          icon={<CloudDownload />}
-          primary
-          label="Primary Text button"
-          onClick={action('onClick')}
-        />
-      </LineStackLayout>
-      <LineStackLayout noMargin>
-        <Text>Buttons with split menus:</Text>
-      </LineStackLayout>
-      <LineStackLayout noMargin>
-        <RaisedButton
-          label="Traditional Raised button"
-          onClick={action('onClick')}
-        />
-        <RaisedButtonWithSplitMenu
-          label="Button with split menu"
-          onClick={action('onClick')}
-          buildMenuTemplate={buildFakeMenuTemplate}
-        />
-        <RaisedButtonWithSplitMenu
-          label="Primary button with split menu"
-          primary
-          onClick={action('onClick')}
-          buildMenuTemplate={buildFakeMenuTemplate}
-        />
-        <RaisedButtonWithSplitMenu
-          label="... and with icon"
-          icon={<Brush />}
-          onClick={action('onClick')}
-          buildMenuTemplate={buildFakeMenuTemplate}
-        />
-        <RaisedButtonWithSplitMenu
-          label="... and disabled"
-          icon={<Brush />}
-          disabled
-          onClick={action('onClick')}
-          buildMenuTemplate={buildFakeMenuTemplate}
-        />
-      </LineStackLayout>
-      <LineStackLayout noMargin>
-        <Text>Buttons with menus:</Text>
-      </LineStackLayout>
-      <LineStackLayout noMargin>
-        <RaisedButton
-          label="Traditional Raised button"
-          onClick={action('onClick')}
-        />
-        <RaisedButtonWithMenu
-          label="Button with menu"
-          buildMenuTemplate={buildFakeMenuTemplate}
-        />
-        <RaisedButtonWithMenu
-          label="... and with icon"
-          icon={<Brush />}
-          buildMenuTemplate={buildFakeMenuTemplate}
-        />
-        <RaisedButtonWithMenu
-          label="... and disabled"
-          icon={<Brush />}
-          disabled
-          buildMenuTemplate={buildFakeMenuTemplate}
-        />
-      </LineStackLayout>
-      <LineStackLayout noMargin>
-        <Text>Icons with menu:</Text>
-      </LineStackLayout>
-      <LineStackLayout noMargin>
-        <ElementWithMenu
-          element={
-            <ToolbarIcon
-              src="res/ribbon_default/bug32.png"
-              tooltip={'ToolbarIcon with menu'}
-            />
-          }
-          buildMenuTemplate={buildFakeMenuTemplate}
-        />
-        <ElementWithMenu
-          element={
-            <IconButton>
-              <FilterList />
-            </IconButton>
-          }
-          buildMenuTemplate={buildFakeMenuTemplate}
-        />
-      </LineStackLayout>
-      <LineStackLayout noMargin>
-        <Text>In a mini toolbar:</Text>
-      </LineStackLayout>
-      <LineStackLayout noMargin>
-        <MiniToolbar>
-          <MiniToolbarText firstChild>Some text:</MiniToolbarText>
-          <IconButton>
-            <Brush />
-          </IconButton>
-          <ElementWithMenu
-            element={
-              <IconButton>
-                <FilterList />
-              </IconButton>
-            }
-            buildMenuTemplate={() => [
-              {
-                label: 'Option 1',
-                click: action('click option 1'),
-              },
-              { type: 'separator' },
-              {
-                label: 'Option 2',
-                click: action('click option 2'),
-              },
-            ]}
-          />
-        </MiniToolbar>
-      </LineStackLayout>
-    </ColumnStackLayout>
-  ));
 
 storiesOf('UI Building Blocks/SelectField', module)
   .addDecorator(muiDecorator)
